HIVE-23016: Extract JdbcConnectionParams from Utils Class
