+++ beeline/src/java/org/apache/hive/beeline/hs2connection/HS2ConnectionFileUtils.java
@@ -23,7 +23,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
-import org.apache.hive.jdbc.Utils.JdbcConnectionParams;
+import org.apache.hive.jdbc.JdbcConnectionParams;
 
 public class HS2ConnectionFileUtils {
