+++ jdbc/src/java/org/apache/hive/jdbc/saml/IJdbcBrowserClientFactory.java
@@ -18,7 +18,7 @@
 
 package org.apache.hive.jdbc.saml;
 
-import org.apache.hive.jdbc.Utils.JdbcConnectionParams;
+import org.apache.hive.jdbc.JdbcConnectionParams;
 import org.apache.hive.jdbc.saml.IJdbcBrowserClient.HiveJdbcBrowserException;
 
 /**
